DSP number 99
This finds the longest palindromic subset of the digits of an integer X.  It does this with a nested for-loop, the outer one moving "forwards" and the inner one moving "backwards" along the integer, storing intermediate results in a separate array.  As such, it requires time O(n^2) and space O(n).